+++ Samples/tvOS-Swift/tvOS-SBSwift/AppDelegate.swift
@@ -13,7 +13,9 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
             // Sampling 100% - In Production you probably want to adjust this
             options.tracesSampleRate = 1.0
 
+            // Experimental features
             options.experimental.enableFileManagerSwizzling = true
+            options.experimental.enableExperimentalViewRenderer = true
         }
         
         return true