+++ Samples/visionOS-Swift/visionOS-Swift/VisionOSSwiftApp.swift
@@ -12,7 +12,10 @@ struct VisionOSSwiftApp: App {
             options.profilesSampleRate = 1.0
             options.attachScreenshot = true
             options.attachViewHierarchy = true
+
+            // Experimental features
             options.experimental.enableFileManagerSwizzling = true
+            options.experimental.enableExperimentalViewRenderer = true
         }
     }