+++ Samples/watchOS-Swift/watchOS-Swift WatchKit Extension/ExtensionDelegate.swift
@@ -9,7 +9,10 @@ class ExtensionDelegate: NSObject, WKExtensionDelegate {
         SentrySDK.start { options in
             options.dsn = "https://6cc9bae94def43cab8444a99e0031c28@o447951.ingest.sentry.io/5428557"
             options.debug = true
+
+            // Experimental features
             options.experimental.enableFileManagerSwizzling = true
+            options.experimental.enableExperimentalViewRenderer = true
         }
         
         SentrySDK.configureScope { scope in