+++ Sources/Sentry/SentrySessionReplayIntegration.m
@@ -68,7 +68,9 @@ static SentryTouchTracker *_touchTracker;
 - (instancetype)initForManualUse:(nonnull SentryOptions *)options
 {
     if (self = [super init]) {
-        [self setupWith:options.sessionReplay enableTouchTracker:options.enableSwizzling];
+        [self setupWith:options.sessionReplay
+                    enableTouchTracker:options.enableSwizzling
+            enableExperimentalRenderer:options.experimental.enableExperimentalViewRenderer];
         [self startWithOptions:options.sessionReplay fullSession:YES];
     }
     return self;
@@ -80,17 +82,28 @@ static SentryTouchTracker *_touchTracker;
         return NO;
     }
 
-    [self setupWith:options.sessionReplay enableTouchTracker:options.enableSwizzling];
+    [self setupWith:options.sessionReplay
+                enableTouchTracker:options.enableSwizzling
+        enableExperimentalRenderer:options.experimental.enableExperimentalViewRenderer];
     return YES;
 }
 
-- (void)setupWith:(SentryReplayOptions *)replayOptions enableTouchTracker:(BOOL)touchTracker
+- (void)setupWith:(SentryReplayOptions *)replayOptions
+            enableTouchTracker:(BOOL)touchTracker
+    enableExperimentalRenderer:(BOOL)enableExperimentalRenderer
 {
     _replayOptions = replayOptions;
     _rateLimits = SentryDependencyContainer.sharedInstance.rateLimits;
-    id<SentryViewRenderer> viewRenderer = [[SentryDefaultViewRenderer alloc] init];
-    _viewPhotographer = [[SentryViewPhotographer alloc] initWithRenderer:viewRenderer
-                                                           redactOptions:replayOptions];
+    id<SentryViewRenderer> viewRenderer;
+    if (enableExperimentalRenderer) {
+        viewRenderer = [[SentryExperimentalViewRenderer alloc] init];
+    } else {
+        viewRenderer = [[SentryDefaultViewRenderer alloc] init];
+    }
+    _viewPhotographer =
+        [[SentryViewPhotographer alloc] initWithRenderer:viewRenderer
+                                           redactOptions:replayOptions
+                               enableExperimentalMasking:enableExperimentalRenderer];
 
     if (touchTracker) {
         _touchTracker = [[SentryTouchTracker alloc]